eulgpt-backend-fastapi/requirements.txt
pydantic-settings-2.9.1 python-dotenv-1.1.0 fastapi-0.115.12